Revert "Merge branch 'Protype-level' into main"
This reverts commit 7f35632596163742c16857faeea1278e8a82eb19, reversing
changes made to f7fc8b75938df5149ceb98568cb51e8d90a541c0.

+++ DH2650/Assets/Scripts/HealthBar.cs
@@ -11,14 +11,14 @@ public class HealthBar : MonoBehaviour
 
     public PlayerHealth playerHealth;
 
-    public void SetHealth(float health)
+    public void SetHealth(int health)
     {
         slider.value = health;
 
         text.text = health.ToString() + "/" + playerHealth.maxHealth.ToString();
     }
 
-    public void SetMaxHealth(float health)
+    public void SetMaxHealth(int health)
     {
         slider.maxValue = health;
         slider.value = health;
